Reorganize file structure in commit lister
