+++ lib/commit_lister/git/get_commits_from_url.rb
@@ -2,7 +2,7 @@ require 'tmpdir'
 require 'net/http'
 require 'directory_utils'
 require_relative './git_wrapper'
-require_relative './commit_constants'
+require_relative '../commit_constants'
 require 'result'
 
 module CommitLister
