+++ lib/commit_lister/validation/validate_url.rb
