+++ lib/commit_lister/validation/validate_url_messages.rb
